cisco-controller: script to gather statistics
Store them in an SQLite3 database.  Use "index" tables and "sighting"
tables.  The "index" tables are the ones that should be relatively
small and not changing (much?) over time -- they list the controllers,
WLANs, APs, and clients (the clients will likely grow slowly over time
with guest devices).

The 2 "sightings" tables will grow quickly over time, and as such, do
not much have much data in them -- they are normalized to point to
entries in the index tables to minimize space usage.

Signed-off-by: Jeff Squyres <jeff@squyres.com>