crontab: move to every 5
Doesn't make sense to update the crontab every minute if we only "git
pull" every 5 minutes.

Also, having the crontab script in the "every1" directory made a bit
of a race condition with the "git pull" script in the "every5"
directory.  I.e., the "git pull" could be running and in the act of
replacing the crontab.txt file when the crontab script runs and
replaces the crontab with the contents of crontab.txt.

In fact, I think that this is what may have happened on Tuesday, Nov
20, 2018, when suddenly there was no crontab any more.

Signed-off-by: Jeff Squyres <jeff@squyres.com>