Script to gather ping-the-gateway data on a continual basis
Simple script to measure some data in Mercy's network.  This script
assumes that there are 1 or more DHCP leases present that it can read.
The script will emit some minimal output to stdout (mainly to indicate
that it's running), and will write most of its metadata to a JSON
file, exiting only when the user hits Ctrl-C.

Signed-off-by: Jeff Squyres <jeff@squyres.com>