gmail-signatures: remove now-unnecessary code
In May 2024, we confirmed that Gmail's WYSIWYG editor is capable of
copy-n-pasting the Mercy stock email signature without any Python
assistance required.  As such, this script is no longer necessary.

Indeed, we actually spent a few hours trying to figure out how to run
this script, but had great difficulty in figuring out domain-wide
delegation google credentials and eventually gave up.  Since there's
no need for this script any more, I'm removing it from the tip of the
branch; it's available in Git history if it's needed.

Signed-off-by: Jeff Squyres <jeff@squyres.com>